+++ Scripts/Controllers/Features/Common/CardCaptainsController.cs
@@ -127,7 +127,7 @@ public class CardCaptainsController : MonoBehaviour
         FindObjectOfType<CurrencyManager>().createCurrency(currencies, currencyPanel);
         currentContent.gameObject.AddComponent<StaggeredSlideAnimation>();
     }
-    public void GetQuantity(int originPrice, object obj, string subType, Transform popupPanel, Transform currencyPanel)
+    public void GetQuantity(double originPrice, object obj, string subType, Transform popupPanel, Transform currencyPanel)
     {
         GameObject quantityObject = Instantiate(quantityPopupPrefab, popupPanel);
