+++ Scripts/Controllers/Features/Common/SpiritBeastController.cs
@@ -137,7 +137,7 @@ public class SpiritBeastController : MonoBehaviour
         FindObjectOfType<CurrencyManager>().createCurrency(currencies, currencyPanel);
         currentContent.gameObject.AddComponent<StaggeredSlideAnimation>();
     }
-    public void GetQuantity(int originPrice, object obj, string subType, Transform popupPanel, Transform currencyPanel)
+    public void GetQuantity(double originPrice, object obj, string subType, Transform popupPanel, Transform currencyPanel)
     {
         GameObject quantityObject = Instantiate(quantityPopupPrefab, popupPanel);
