+++ Scripts/Controllers/Features/User/UserCardLifeController.cs
@@ -184,7 +184,7 @@ public class UserCardLifeController : MonoBehaviour
                 AudioManager.Instance.PlaySFX(AudioConstants.SFX.BUTTON_CLICK);
                 CardLives currentCard = new CardLives();
                 currentCard = UserCardLifeService.Create().GetUserCardLifeById(User.CurrentUserId, cardLife.Id);
-                int totalExperiment = currentCard.Experiment;
+                double totalExperiment = currentCard.Experiment;
                 int currentLevel = currentCard.Level;
                 int experimentCondition = currentLevel == 0 ? 100 : currentLevel * 100;
                 int userMaxLevel = User.CurrentUserLevel;
@@ -210,7 +210,7 @@ public class UserCardLifeController : MonoBehaviour
             {
                 AudioManager.Instance.PlaySFX(AudioConstants.SFX.BUTTON_CLICK);
                 CardLives currentCard = UserCardLifeService.Create().GetUserCardLifeById(User.CurrentUserId, cardLife.Id);
-                int totalExperiment = currentCard.Experiment;
+                double totalExperiment = currentCard.Experiment;
                 int currentLevel = currentCard.Level;
                 int originalLevel = currentLevel;
                 int experimentCondition = currentLevel == 0 ? 100 : currentLevel * 100;
@@ -290,8 +290,8 @@ public class UserCardLifeController : MonoBehaviour
             breakthroughButton.onClick.AddListener(() =>
             {
                 AudioManager.Instance.PlaySFX(AudioConstants.SFX.BUTTON_CLICK);
-                int requiredQuantity = cardLife.Star + 1;
-                int totalItemQuantity = 0;
+                double requiredQuantity = cardLife.Star + 1;
+                double totalItemQuantity = 0;
 
                 // Kiểm tra số lượng vòng phép
                 bool hasEnoughMagicFormation = cardLife.Quantity >= requiredQuantity;
@@ -313,7 +313,7 @@ public class UserCardLifeController : MonoBehaviour
                     else
                     {
                         // Nếu vòng phép không đủ, dùng cả vòng phép + vật phẩm để bù vào
-                        int remainingRequired = requiredQuantity - cardLife.Quantity;
+                        double remainingRequired = requiredQuantity - cardLife.Quantity;
                         cardLife.Quantity = 0; // Dùng hết vòng phép
 
                         foreach (Items items1 in items)
